LPCPLAY modifications: Added pitch data preprocessing options to better handle ugly pitch analysis data.  Needed to move some source files around to do this. Changed algorithm used to generate crossfaded noise/voiced output (what you get when you use low and hi thresh numbers).  Fixed some bugs in the embedded version of the setup code.

+++ insts/std/LPCPLAY/LPCPLAY.h
@@ -89,7 +89,7 @@ private:
 	float	_risetime, _decaytime;				// enveloping
 
 	// These are set and used within LPCPLAY.
-    double  _actualWeight;
+    float   _actualWeight;
 	double	_pitch;
 	double	_transposition;
     double  _voicedFrameIncrement;
